% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTau.R
\name{calcTau}
\alias{calcTau}
\title{Calculate Mixture Variance}
\usage{
calcTau(alpha, sigma, truncation)
}
\arguments{
\item{alpha}{Significance level}

\item{sigma}{Population standard deviation}

\item{truncation}{Desired truncation time for mSPRT}
}
\value{
tau Optimal mixture variance \eqn{\tau} for mSPRT.
}
\description{
Calculate Mixture Variance
}
\references{
\emph{Johari, R., Koomen, P., Pekelis, L. & Walsh, D. 2017, "Peeking at A/B Tests: Why it matters, and what to do about it", ACM, , pp. 1517}
}
