% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSPRT.default.R
\name{mSPRT.default}
\alias{mSPRT.default}
\title{Perform mixture Sequential Probability Ratio Test}
\arguments{
\item{x, y}{Numeric vectors}

\item{xpre, ypre}{Numeric vectors of pre-experiment data}

\item{sigma}{Population standard deviation}

\item{tau}{Mixture variance}

\item{theta}{Hypothesised difference between \code{x} and \code{y}}

\item{distribution}{The desired distribution. Currently, only \code{normal} is implemented.}

\item{alpha}{Significance level}
}
\value{
The likelihood ratio
}
\description{
Perform mixture Sequential Probability Ratio Test
}
\references{
Johari, R., Koomen, P., Pekelis, L. & Walsh, D. 2017, "Peeking at A/B Tests: Why it matters, and what to do about it", ACM, , pp. 1517
}
