% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mSPRT.R
\name{plot.mSPRT}
\alias{plot.mSPRT}
\title{plot.mSPRT}
\usage{
\method{plot}{mSPRT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mSPRT}}

\item{...}{Further arguments}
}
\description{
plot.mSPRT
}
