% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mSPRT.R
\name{print.mSPRT}
\alias{print.mSPRT}
\title{print.mSPRT}
\usage{
\method{print}{mSPRT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mSPRT}}

\item{...}{Further arguments}
}
\description{
print.mSPRT
}
