% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testing}
\alias{testing}
\title{Calculate mixture Sequential Probability Ratio Test in Cpp}
\arguments{
\item{x, y}{Numeric vectors}

\item{sigma}{Population standard deviation}

\item{tau}{Mixture variance}

\item{theta}{Hypothesised difference between \code{x} and \code{y}}

\item{distribution}{The desired distribution.}

\item{alpha}{Significance level}
}
\value{
The likelihood ratio
}
\description{
Calculate mixture Sequential Probability Ratio Test in Cpp
}
