% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSPRT.R
\name{mSPRT}
\alias{mSPRT}
\title{Perform mixture Sequential Probability Ratio Test}
\usage{
mSPRT(x, y, sigma, tau, theta = 0, distribution = "normal",
  alpha = 0.05)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{sigma}{Population standard deviation}

\item{tau}{Mixture variance}

\item{theta}{Hypothesised difference between \code{x} and \code{y}}

\item{alpha}{Significance level}
}
\value{
The likelihood ratio
}
\description{
Perform mixture Sequential Probability Ratio Test
}
