% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSPRT.R
\name{mSPRT}
\alias{mSPRT}
\title{Calculate mixture Sequential Probability Ratio Test}
\usage{
mSPRT(x, y, sigma, tau, theta = 0, distribution = "normal",
  alpha = 0.05, useCpp = F)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{sigma}{Population standard deviation}

\item{tau}{Mixture variance}

\item{theta}{Hypothesised difference between \code{x} and \code{y}}

\item{distribution}{The desired distribution.}

\item{alpha}{Significance level}

\item{useCpp}{Boolean. Use C++ for calculations? Useful for running many tests as it reduces runtime substantially}
}
\value{
The likelihood ratio
}
\description{
Calculate mixture Sequential Probability Ratio Test
}
\section{Details}{

With normal data and normal prior, the closed form solution of the probability ratio after \eqn{n} observations have been collected is:
\deqn{\tilde{\Lambda}_n = \sqrt{\frac{2\sigma^2}{V_n + n\tau^2}}\exp{\left(\frac{n^2\tau^2(\bar{Y}_n - \bar{X}_n-\theta_0)^2}{4\sigma2(2\sigma^2+n\tau^2)}\right)}.}
With Bernoulli data, the closed form solution is:
\deqn{\tilde{\Lambda}_n = \sqrt{\frac{V_n}{V_n + n\tau^2}}\exp{\left(\frac{n^2\tau^2(\bar{Y}_n - \bar{X}_n-\theta_0)^2}{2V_n(V_n+n\tau^2)}\right)}.}
}

\references{
\emph{Johari, R., Koomen, P., Pekelis, L. & Walsh, D. 2017, 'Peeking at A/B Tests: Why it matters, and what to do about it', ACM, , pp. 1517}
}
