% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSPRT.R
\name{mSPRT}
\alias{mSPRT}
\title{Calculate mixture Sequential Probability Ratio Test}
\usage{
mSPRT(x, y, sigma, tau, theta = 0, distribution = "normal",
  alpha = 0.05)
}
\arguments{
\item{x, y}{Numeric vectors}

\item{sigma}{Population standard deviation}

\item{tau}{Mixture variance}

\item{theta}{Hypothesised difference between \code{x} and \code{y}}

\item{distribution}{The desired distribution. Currently, only \code{normal} is implemented.}

\item{alpha}{Significance level}
}
\value{
The likelihood ratio
}
\description{
Calculate mixture Sequential Probability Ratio Test
}
\references{
\emph{Johari, R., Koomen, P., Pekelis, L. & Walsh, D. 2017, "Peeking at A/B Tests: Why it matters, and what to do about it", ACM, , pp. 1517}
}
